cast for datatype of score.
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@3488 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/switch_ivr.c
@@ -496,7 +496,7 @@ SWITCH_DECLARE(switch_status_t) switch_ivr_record_file(switch_core_session_t *se
                 j += read_codec->implementation->number_of_channels;
             }
 		
-            score = energy / samples;
+            score = (uint32_t)(energy / samples);
             if (score < fh->thresh) {
                 if (!--fh->silence_hits) {
                     break;
